Add project maturity badge
This commit removes the advisory text in the README in favor of a new
shiny badge which indicates production readiness.

In sending this, I noticed that GH actions are running on this
repository. Nice job! Would be great to add a badge here for the
meaningful action runs as well.

Signed-off-by: Evan Gilman <egilman@vmware.com>